Update / revision
I added more "widgets" to the system. I'll probably add more for tomorrow but I need to rest for now bc I want to. The images that are supposedly to show up in the system are not yet to be seen since I'll plan to look for those tomorrow and resize it until it looked amazing to the interface.